requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1

# For PostgreSQL (production)
psycopg2-binary==2.9.9

# Africa's Talking SDK
africastalking==1.2.6

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & Forms
python-multipart==0.0.6
httpx==0.25.2

# Templates (for simple dashboard)
jinja2==3.1.2

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
